#10071: Merge log_bw to TTNN
