+++ docs/source/ttnn/ttnn/dependencies/tt_lib.rst
@@ -814,7 +814,7 @@ Backward Operations
 
 .. autofunction:: tt_lib.tensor.fill_bw
 
-.. autofunction:: tt_lib.tensor.log_bw
+.. autofunction:: tt_lib.tensor.unary_sub_bw
 
 .. autofunction:: tt_lib.tensor.abs_bw
