+++ docs/source/ttnn/ttnn/ttnn/floor_bw.rst
@@ -1,0 +1,6 @@
+.. _ttnn.floor_bw:
+
+ ttnn.floor_bw
+ ##############
+
+ .. autofunction:: ttnn.floor_bw
