+++ docs/source/ttnn/ttnn/ttnn/log_bw.rst
@@ -1,0 +1,6 @@
+.. _ttnn.log_bw:
+
+ ttnn.log_bw
+ ############
+
+ .. autofunction:: ttnn.log_bw
