+++ docs/source/ttnn/ttnn/ttnn/round_bw.rst
@@ -1,0 +1,6 @@
+.. _ttnn.round_bw:
+
+ ttnn.round_bw
+ ##############
+
+ .. autofunction:: ttnn.round_bw
