+++ tt_eager/tt_dnn/op_library/backward/backward_ops.hpp
@@ -101,7 +101,7 @@ std::vector<std::optional<Tensor>> tanh_bw(
 std::vector<Tensor> fill_bw(
     const Tensor& grad, const MemoryConfig& output_mem_config = operation::DEFAULT_OUTPUT_MEMORY_CONFIG);
 
-std::vector<Tensor> log_bw(
+std::vector<Tensor> unary_sub_bw(
     const Tensor& grad,
     const Tensor& input,
     const MemoryConfig& output_mem_config = operation::DEFAULT_OUTPUT_MEMORY_CONFIG);
